service-posts/requirements.txt
grpcio
sqlalchemy
psycopg2-binary
grpcio-tools
